refactor: let httpRequest method use fetch options
